+++ packages/core/client/src/schema-component/antd/action/Action.Designer.tsx
@@ -1,6 +1,7 @@
 import { ISchema, useField, useFieldSchema } from '@formily/react';
 import { isValid } from '@formily/shared';
-import React from 'react';
+import { uid } from '@nocobase/utils';
+import React, { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useActionContext, useCompile, useDesignable } from '../..';
 import { GeneralSchemaDesigner, SchemaSettings } from '../../../schema-settings';
@@ -13,7 +14,67 @@ export const ActionDesigner = (props) => {
   const compile = useCompile();
   const isPopupAction = ['create', 'update', 'view', 'customizePopup'].includes(fieldSchema['x-action'] || '');
   const context = useActionContext();
+  const [initialSchema, setInitialSchema] = useState<{ uid: string; schema: ISchema }>();
 
+  useEffect(() => {
+    const schemaUid = uid();
+    const schema: ISchema = {
+      type: 'void',
+      'x-uid': schemaUid,
+      properties: {
+        modal: {
+          'x-component': 'Action.Modal',
+          'x-component-props': {
+            width: 520,
+          },
+          type: 'void',
+          title: '{{ t("Assigned field value") }}',
+          properties: {
+            tip: {
+              type: 'void',
+              'x-editable': false,
+              'x-decorator': 'FormItem',
+              'x-component': 'Markdown.Void',
+              'x-index': 0,
+              'x-component-props': {
+                content: t('Save assigned field value after click button'),
+              },
+            },
+            grid: {
+              type: 'void',
+              'x-component': 'Grid',
+              'x-initializer': 'CustomFormItemInitializers',
+              properties: {},
+            },
+            footer: {
+              'x-component': 'Action.Modal.Footer',
+              type: 'void',
+              properties: {
+                cancel: {
+                  type: 'void',
+                  title: '{{ t("Cancel") }}',
+                  'x-component': 'Action',
+                  'x-component-props': {
+                    useAction: '{{ useCancelAction }}',
+                  },
+                },
+                submit: {
+                  type: 'void',
+                  title: '{{ t("Submit") }}',
+                  'x-component': 'Action',
+                  'x-component-props': {
+                    type: 'primary',
+                    useAction: '{{ useSubmitAction }}',
+                  },
+                },
+              },
+            },
+          },
+        },
+      },
+    };
+    setInitialSchema({ uid: schemaUid, schema });
+  }, []);
   return (
     <GeneralSchemaDesigner {...props} disableInitializer>
       <SchemaSettings.ModalItem
@@ -87,6 +148,7 @@ export const ActionDesigner = (props) => {
       {isValid(fieldSchema?.['x-action-settings']?.assignedValues) && (
         <SchemaSettings.ActionModalItem
           title={t('Assigned field value')}
+          initialSchema={initialSchema}
           initialValues={fieldSchema?.['x-action-settings']?.assignedValues}
           uid={fieldSchema?.['x-action-settings']?.schemaUid}
           onSubmit={(assignedValues) => {
